Changed dependency direction to ClientService -> ManagerService
