+++ timesheet_backend/src/main/java/com/g4l/timesheet_backend/models/requests/ClientTeamRequest.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Data
 public class ClientTeamRequest {
-    public String clientTeamId;
+    public String id;
     public String teamName;
     public String clientId;
     public String managerId;
